regen(go-src-converted): rebank the full stdlib with the 07-12 converter fixes
The committed full-conversion stdlib (src/go-src-converted) was last bulk-converted
at the 07-10 "302/302 compiles" milestone (5cc07f8f3), BEFORE the 07-12 Phase-4
converter fixes landed. Because `deploy-core stdlib` ships the COMMITTED tree, a
clean-repo deploy + the README fatih/color steps still hit stale-stdlib bugs — most
visibly fmt's `ppFree.Put(p)` boxing the pp VALUE, which panics on the second fmt
call ("interface conversion: interface {} is pp, not ж<pp>"). Regenerating banks the
current converter's output into the committed tree so operational milestones (the
fatih/color run) reproduce from a clean checkout with no manual fmt refresh.

Regenerated all 301 auto-converted packages with the current converter (worktree
HEAD 1a005ff64) — a full `-stdlib` reconvert into a clean tree, overlaid onto
go-src-converted with the hand-owned files preserved. Banks, notably:
  * pp-pool pointer-value -> `any`-arg boxing (79ef00e9d): fmt now emits
    `ᏑppFree.Put(Ꮡp)`, and the same box-the-pointer fix lands across go/types,
    net, os, math/big, database/sql, ... (a `*T` value passed whole to `...any`
    now boxes the pointer `Ꮡx`, matching Go).
  * untyped-int-const -> nint boxing (96b13eec2).
  * switch-fallthrough-default (5fce5f0f1): fmt's printValue default arm no longer
    fires after a matched non-fallthrough case — the real blocker for composite %v.

Footprint: 86 .cs with real content changes across ~30 packages. The 301 .csproj
show as touched but are line-ending-only (autocrlf) — zero staged content change.
The 19 [module:GoManualConversion] whole-files (runtime2, sync natives, syscall
dll_windows, reflect/abi Type+Value bridge, unsafe, ...) and 11 *_impl.cs
companions are restored verbatim from HEAD — the converter's fresh output for them
is discarded.

Verified: all 302 packages compile clean (go-src-converted.slnx, 0 CS errors);
behavioral check-no-regression byte-identical (converter source unchanged).

This is a large generated commit by design.

+++ src/go-src-converted/crypto/x509/parser.cs
@@ -918,7 +918,7 @@ internal static (ж<Certificate>, error) parseCertificate(slice<byte> der) {
     if (!tbs.ReadASN1(Ꮡtbs, cryptobyte_asn1.SEQUENCE)) {
         return (default!, errors.New("x509: malformed tbs certificate"u8));
     }
-    if (!tbs.ReadOptionalASN1Integer(cert.of(Certificate.ᏑVersion), ((cryptobyte_asn1.Tag)0).Constructed().ContextSpecific(), 0)) {
+    if (!tbs.ReadOptionalASN1Integer(cert.of(Certificate.ᏑVersion), ((cryptobyte_asn1.Tag)0).Constructed().ContextSpecific(), (nint)(0))) {
         return (default!, errors.New("x509: malformed version"u8));
     }
     if ((~cert).Version < 0) {
